feat: cards UI done
